+++ lib/ooxml_parser/xlsx_parser/xlsx_data/view_model/workbook/worksheet/xlsx_row/xlsx_cell.rb
@@ -19,33 +19,29 @@ module OoxmlParser
     # @return [XlsxCell] result of parsing
     def parse(node)
       text_string_id = nil
-      text_string_id = node.attribute('s').value unless node.attribute('s').nil?
+      text_string_id = node.attribute('s').value if node.attribute('s')
       @style = CellStyle.new(parent: self).parse(text_string_id)
-      if node.attribute('t').nil?
-        @raw_text = node.xpath('xmlns:v').text
+      if node.attribute('t')
+        node.attribute('t').value == 's' ? get_shared_string(node.xpath('xmlns:v').text) : @raw_text = node.xpath('xmlns:v').text
       else
-        node.attribute('t').value == 's' ? XlsxCell.get_shared_string(node.xpath('xmlns:v').text, self) : @raw_text = node.xpath('xmlns:v').text
+        @raw_text = node.xpath('xmlns:v').text
       end
       @formula = node.xpath('xmlns:f').text unless node.xpath('xmlns:f').text == ''
-      @text = @raw_text.dup unless @raw_text.nil?
+      @text = @raw_text.dup if @raw_text
       @text.insert(0, "'") if @style.quote_prefix
       self
     end
 
+    private
+
     # Get shared string by it's number
     # @param [String] value number of shared string
-    # @param [XlsxCell] cell to write value of string
     # @return [Nothing]
-    def self.get_shared_string(value, cell)
+    def get_shared_string(value)
       return '' if value == ''
-      XLSXWorkbook.shared_strings[value.to_i].xpath('*').each do |si_node_child|
-        case si_node_child.name
-        when 'r'
-          cell.character = ParagraphRun.new(parent: cell).parse(si_node_child)
-        when 't'
-          cell.raw_text = si_node_child.text
-        end
-      end
+      string = root_object.shared_strings_table.string_indexes[value.to_i]
+      @character = string.run
+      @raw_text = string.text
     end
   end
 end